Add `filter` operation
Filtering is similar to the old `retain` operation but creates a new store instead of operating in-place, as it is _significantly_ easier to maintain durability requirements.

Also adds a `cleanup` operation to the hash bridge that rolls up the FSTs into one, and uses the stream of values to `filter` the lookup.